UDFs for comparator and math expression
