+++ server/src/main/java/io/druid/initialization/Initialization.java
@@ -30,7 +30,6 @@ import com.google.inject.Injector;
 import com.google.inject.Key;
 import com.google.inject.Module;
 import com.google.inject.util.Modules;
-
 import io.druid.curator.CuratorModule;
 import io.druid.curator.discovery.DiscoveryModule;
 import io.druid.guice.AWSModule;
@@ -58,6 +57,7 @@ import io.druid.guice.annotations.Json;
 import io.druid.guice.annotations.Smile;
 import io.druid.guice.http.HttpClientModule;
 import io.druid.guice.security.DruidAuthModule;
+import io.druid.jackson.FunctionInitializer;
 import io.druid.java.util.common.ISE;
 import io.druid.java.util.common.logger.Logger;
 import io.druid.metadata.storage.derby.DerbyMetadataStorageDruidModule;
@@ -364,7 +364,12 @@ public class Initialization
       extensionModules.addModule(module);
     }
 
-    return Guice.createInjector(Modules.override(intermediateModules).with(extensionModules.getModules()));
+    return Guice.createInjector(
+        Modules.combine(
+            Modules.override(intermediateModules).with(extensionModules.getModules()),
+            new FunctionInitializer()
+        )
+    );
   }
 
   private static class ModuleList
